+++ backend/controllers/userController.js
@@ -6,7 +6,7 @@ export const getMe = async (req, res) => {
     if (!user) {
       return res.status(404).json({ message: "User not found" });
     }
-    res.json({ username: user.username, email: user.email });
+    res.json({ _id: user._id, username: user.username, email: user.email });
   } catch (err) {
     res.status(500).json({ message: "Server error" });
   }
